Move audio clips DB into a separate file
